# backend/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Cadrly API
  version: 0.1.0
  description: Приложение для обмена коллекционными карточками
servers:
  - description: Test server
    url: http://localhost:8080/api
...
paths:
...
  /home/search:
    get:
      tags: [Home]
      summary: Поиск пользователей
      description: |
        Поиск пользователей по никнейму или ID
      parameters:
        - in: query
          name: query
          description: Никнейм или ID пользователя
          schema:
            type: string
      responses:
        200:
          description: Список найденных пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        400:
          description: Неверный запрос
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Введите минимум 3 символа для поиска'
        404:
          description: Пользователь не найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Пользователь с введеными данными не найден'
        500:
          $ref: '#/components/responses/ServerError'
  /home/notifications:
    get:
      tags: [Home]
      summary: Просмотр уведомлений
      description: |
        Возвращает список уведомлений пользователя
      security:
        - bearerAuth: []
      responses:
        200:
          description: Список уведомлений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/ServerError'
  /home/notifications/{notification_ID}:
    get:
      tags: [Home]
      summary: Просмотр уведомления
      description: |
        Возвращает информацию по конкретному уведомлению
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: notification_ID
          required: true
          schema:
            type: integer
            example: 123
            description: 'ID уведомления'
      responses:
        200:
          description: Детальная информация об уведомлении
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/ServerError'
  /home/notifications/{notification_ID}/navigate:
    post:
      tags: [Home]
      summary: Переход по ссылке в  уведомлении
      description: |
        Перенаправляет на страницу обмена, связанного с уведомлением
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: notification_ID
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Перенаправление на страницу обмена
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirectUrl:
                    type: string
                    example: '/trades/123'
        404:
          description: Предложение обмена не найдено
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Данное предложение обмена уже не существует'
        500:
          $ref: '#/components/responses/ServerError'
  /home/generate-card/themes:
    get:
      tags: [Home]
      summary: Получение списка тем для генерации карт
      description: |
        Возвращает список доступных тем для генерации уникальных карт
      security:
        - bearerAuth: []
      responses:
        200:
          description: Список тем
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Theme'
        500:
          $ref: '#/components/responses/ServerError'
  /home/generate-card:
    post:
      tags: [Home]
      summary: Генерация уникальной карты
      description: |
        Генерирует уникальную карту по выбранной теме, если на балансе достаточно средств
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                theme_ID:
                  type: integer
                  example: 901
                  description: 'ID выбранной темы'
              required: [theme_ID]
      responses:
        200:
          description: Карта успешно сгенерирована
          content:
            application/json:
              schema:
                type: object
                properties:
                  card:
                    $ref: '#/components/schemas/Card'
                  newBalance:
                    type: integer
                    example: 1500
        402:
          description: Недостаточно средств
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Недостаточно средств для генерации карты'
                  requiredAmount:
                    type: integer
                    example: 1000
        500:
          $ref: '#/components/responses/ServerError'
  /home/news:
    get:
      tags: [Home]
      summary: Получение списока новостей
      description: |
        Возвращает список новостей
      responses:
        200:
          description: Список новостей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/News'
        404:
          description: Новостей не найдено
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Новостей пока нет'
        500:
          $ref: '#/components/responses/ServerError'
  /home/news/{news_ID}:
    get:
      tags: [Home]
      summary: Просмотр новости
      description: |
        Возвращает полную информацию по конкретной новости
      parameters:
        - in: path
          name: news_ID
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Детальная информация о новости
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/News'
        500:
          $ref: '#/components/responses/ServerError'
  /home/quests:
    get:
      tags: [Home]
      summary: Получение списка квестов
      description: |
        Возвращает список ежедневных и еженедельных квестов с их статусом
      security:
        - bearerAuth: []
      responses:
        200:
          description: Список квестов
          content:
            application/json:
              schema:
                type: object
                properties:
                  dailyQuests:
                    type: array
                    items:
                      $ref: '#/components/schemas/Quest'
                  weeklyQuests:
                    type: array
                    items:
                      $ref: '#/components/schemas/Quest'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/ServerError'
  /home/quests/{quest_ID}/change-status:
    post:
      tags: [Home]
      summary: Изменение статуса выполнения квеста
      description: |
        Изменение статуса выполнения квеста
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: quest_ID
          required: true
          schema:
            type: integer
            example: 123
            description: 'ID квеста'
      responses:
        200:
          description: Статус квеста успешно изменен
          content:
            application/json:
              schema:
                type: object
                properties:
                  quest:
                    $ref: '#/components/schemas/Quest'
                  message:
                    type: string
                    example: 'Статус квеста успешно изменен'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/ServerError'
  /home/quests/claim-reward:
    post:
      tags: [Home]
      summary: Получение награды за выполнение квестов
      description: |
        Выдает награду за выполнение квестов
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                questType:
                  type: string
                  enum: [daily, weekly]
                  example: 'daily'
              required: [questType]
      responses:
        200:
          description: Награда получена
          content:
            application/json:
              schema:
                type: object
                properties:
                  receivedCoins:
                    type: integer
                    example: 500
                  receivedPacks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pack'
                  newBalance:
                    type: integer
                    example: 2000
        400:
          description: Не все квесты выполнены
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Выполнены не  все квесты, необходимые для получения награды'
        500:
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
    News:
      type: object
      properties:
        news_ID:
          type: integer
          example: 1
          description: 'Уникальный идентификатор новости'
        title:
          type: string
          example: 'Новое обновление системы'
          description: 'Заголовок новости'
        content:
          type: string
          example: 'Мы добавили новые карты в коллекцию'
          description: 'Содержание новости'
        pictures:
          type: array
          items:
            type: string
            format: url
          example:
            ['https://example.com/news1.jpg', 'https://example.com/news2.jpg']
          description: 'Ссылки на изображения'
        datePosted:
          type: string
          format: date-time
          example: '2024-05-20T14:48:00Z'
          description: 'Дата публикации'
      required:
        - news_ID
        - title
        - content
        - datePosted
    Achievement:
      type: object
      properties:
        achievement_ID:
          type: integer
          example: 101
          description: 'Уникальный идентификатор достижения'
        name:
          type: string
          example: 'Мастер коллекционирования'
          description: 'Название достижения'
        imageURL:
          type: string
          format: url
          example: 'https://example.com/achievements/master.png'
          description: 'Ссылка на изображение'
        description:
          type: string
          example: 'Соберите 10 полных коллекций'
          description: 'Описание достижения'
        isUnlocked:
          type: boolean
          example: false
          description: 'Статус получения'
      required:
        - achievement_ID
        - name
        - imageURL
        - description
        - isUnlocked
    Notification:
      type: object
      properties:
        notification_ID:
          type: integer
          example: 201
          description: 'Уникальный идентификатор уведомления'
        user:
          $ref: '#/components/schemas/User'
          description: 'Пользователь, которому отправлено уведомление'
        message:
          type: string
          example: 'Ваша карта была успешно обменяна'
          description: 'Текст уведомления'
        links:
          type: array
          items:
            type: string
            format: url
          example: ['https://example.com/exchange-details']
          description: 'Ссылка на подробности обмена'
        notificationDateTime:
          type: string
          format: date-time
          example: '2024-05-20T15:30:00Z'
          description: 'Дата и время уведомления'
      required:
        - notification_ID
        - user
        - message
        - notificationDateTime
...
    Pack:
      type: object
      properties:
        pack_ID:
          type: integer
          example: 401
          description: 'Уникальный идентификатор набора'
        name:
          type: string
          example: 'Стартовый набор'
          description: 'Название набора'
        imageURL:
          type: string
          format: url
          example: 'https://example.com/packs/starter.png'
          description: 'Ссылка на изображение'
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Карты в наборе'
        price:
          type: integer
          example: 1000
          description: 'Цена'
      required:
        - pack_ID
        - name
        - imageURL
        - cards
        - price
    Card:
      type: object
      properties:
        card_ID:
          type: integer
          example: 501
          description: 'Уникальный идентификатор карты'
        name:
          type: string
          example: 'Ледяной феникс'
          description: 'Название карты'
        imageURL:
          type: string
          format: url
          example: 'https://example.com/cards/pheonix.png'
          description: 'Ссылка на изображение'
        description:
          type: string
          example: 'Ледяной Феникс — это мифическое существо, воплощающее силу зимы и вечного обновления. Его перья сверкают как морозный утренний иней, а глаза светятся холодным синим светом.'
          description: 'Описание карты'
        rarity:
          type: string
          enum: ['Обычная', 'Редкая', 'Эпическая', 'Легендарная', 'Уникальная']
          example: 'Редкая'
          description: 'Редкость карты'
        min_price:
          type: integer
          example: 50
          description: 'Минимальная цена - используется для разбора карточки'
        theme:
          type: string
          nullable: true
          example: 'Мифическое существо'
          description: 'Тематика, на которую была сгенерирвана карта (только для уникальных карт)'
        isGenerated:
          type: boolean
          example: false
          description: 'Флаг показывающий сгенерированная ли карта (true для уникальных, false для остальных)'
      required:
        - card_ID
        - name
        - imageURL
        - rarity
        - min_price
        - isGenerated
    User:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
          description: 'Уникальный идентификатор'
        username:
          type: string
          example: 'Коллекционер_123'
          description: 'Имя пользователя'
        email:
          type: string
          format: email
          example: 'user@example.com'
          description: 'Электронная почта'
        password:
          type: string
          format: password
          writeOnly: true
          description: 'Пароль'
        favoriteCards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Избранные карты'
        onChange:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Карты на обмен'
        inventoryCards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Карты в инвентаре'
        balance:
          type: integer
          example: 2500
          description: 'Баланс'
        avatar_url:
          type: string
          format: url
          example: 'https://example.com/avatars/user1.png'
          description: 'Аватар пользователя'
        achievements:
          type: array
          items:
            $ref: '#/components/schemas/Achievement'
          description: 'Достижения'
        favoriteAchievements:
          type: array
          items:
            $ref: '#/components/schemas/Achievement'
          description: 'Избранные достижения'
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
          description: 'Уведомления'
      required:
        - user_id
        - username
        - email
        - password
        - balance
        - inventoryCards
        - achievements
...
    Theme:
      type: object
      properties:
        theme_ID:
          type: integer
          example: 901
          description: 'Уникальный идентификатор темы'
        name:
          type: string
          example: 'Мифические существа'
          description: 'Название темы'
        description:
          type: string
          example: 'Создайте изображение мифического существа'
          description: 'Описание темы'
      required:
        - theme_ID
        - name
...
    Quest:
      type: object
      properties:
        quest_ID:
          type: integer
          example: 1
          description: 'Уникальный идентификатор квеста'
        name:
          type: string
          example: 'Соберите 5 карт'
          description: 'Название квеста'
        description:
          type: string
          example: 'Соберите 5 карт из коллекции Драконы'
          description: 'Описание квеста'
        progress:
          type: integer
          example: 3
          description: 'Текущий прогресс'
        target:
          type: integer
          example: 5
          description: 'Целевое значение'
        rewardCoins:
          type: integer
          example: 200
          description: 'Награда в монетах'
        rewardPacks:
          type: array
          items:
            $ref: '#/components/schemas/Pack'
          description: 'Награда в наборах'
        isCompleted:
          type: boolean
          example: false
          description: 'Выполнен ли квест'
        isClaimed:
          type: boolean
          example: false
          description: 'Получена ли награда'
      required:
        - quest_ID
        - name
        - description
        - progress
        - target
        - rewardCoins
        - isCompleted
        - isClaimed
  responses:
...
    Unauthorized:
      description: Ошибка аутентификации
    ServerError:
      description: Что-то пошло не так
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Что-то пошло не так, попробуйте позже'
              code:
                type: string
                example: 'INTERNAL_SERVER_ERROR'
              timestamp:
                type: string
                format: date-time
                example: '2024-05-20T12:00:00Z'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT